+++ CHANGES.md
@@ -2,6 +2,7 @@ master (in development)
 =======================
 
 - Add `flycheck-pos-tip-timeout` [GH-9]
+- Add `flycheck-pos-tip-hide-function`
 
 0.1 (Nov 16, 2015)
 ==================